test: isolate child namespace harness
